feat: add Repository and feat docker-compose.yml

+++ stats-service/server/src/main/java/ru/practicum/model/EndpointHit.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.boot.autoconfigure.batch.BatchDataSource;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
@@ -21,7 +20,7 @@ public class EndpointHit {
     @Id
     @Column(name = "id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "app")
     private String app;
     @Column(name = "uri")
